refactor: consolidate world tests into part libraries (#3843)
Merge fragmented fog, connectivity, province lookup, and capital test
files into single entry points with part files, cutting test entry count
from 64 to 51 (≥25% reduction vs 76 baseline) while respecting the
400-line domain test cap.
