+++ packages/colonizethis_world/test/world/capital_and_gp_fall_reassignment_part.dart
@@ -1,46 +1,6 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_world/src/world/capital_and_gp_fall.dart';
-import 'package:colonizethis_world/src/world/capital_reassignment_fatal.dart';
-import 'package:colonizethis_world/src/world/province_lookup.dart';
-import 'package:colonizethis_test/test.dart';
+part of 'capital_test.dart';
 
-/// Coverage uplift for `colonizethis_world` (Refs #3290 Phase 1 follow-up).
-///
-/// Exercises runtime capital reassignment after combat for Great Powers,
-/// Minor Nations, and Tribes in `lib/src/world/capital_and_gp_fall_reassignment.dart`.
-/// SPEC/game/capital-and-connectivity § Capital loss and reassignment.
-const _emptyTopology = MapTopology();
-
-CapitalTile _tile(String provinceId, int x, int y) => CapitalTile(
-  regionId: ProvinceId.regionIdFrom(provinceId),
-  provinceId: provinceId,
-  x: x,
-  y: y,
-);
-
-Game _gpGame({
-  required List<Province> provinces,
-  String capitalProvinceId = 'oldWorld|cap',
-}) => Game(
-  id: 'g-gp-reassign',
-  worldState: WorldState(
-    turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-    oldWorld: RegionData(provinces: provinces),
-    newWorld: const RegionData(),
-  ),
-  players: [
-    Player(
-      id: 'p1',
-      displayName: 'P1',
-      isHuman: true,
-      capitalProvinceId: capitalProvinceId,
-      capitalTile: _tile(capitalProvinceId, 0, 0),
-    ),
-  ],
-);
-
-void main() {
+void _capital_and_gp_fall_reassignment_testTests() {
   group('applyCapitalReassignmentAfterCombat (Great Power)', () {
     test('reassigns capital to remaining owned province after loss', () {
       final game = _gpGame(
@@ -166,7 +126,7 @@ void main() {
           MinorNation(
             id: 'm1',
             capitalProvinceId: 'oldWorld|mcap',
-            capitalTile: _tile('oldWorld|mcap', 0, 0),
+            capitalTile: _tile('oldWorld|mcap', x: 0, y: 0),
           ),
         ],
       );
@@ -198,7 +158,7 @@ void main() {
           MinorNation(
             id: 'm1',
             capitalProvinceId: 'oldWorld|mcap',
-            capitalTile: _tile('oldWorld|mcap', 0, 0),
+            capitalTile: _tile('oldWorld|mcap', x: 0, y: 0),
           ),
         ],
       );
@@ -234,7 +194,7 @@ void main() {
           Tribe(
             id: 't1',
             capitalProvinceId: 'newWorld|tcap',
-            capitalTile: _tile('newWorld|tcap', 0, 0),
+            capitalTile: _tile('newWorld|tcap', x: 0, y: 0),
           ),
         ],
       );
@@ -264,7 +224,7 @@ void main() {
           Tribe(
             id: 't1',
             capitalProvinceId: 'newWorld|tcap',
-            capitalTile: _tile('newWorld|tcap', 0, 0),
+            capitalTile: _tile('newWorld|tcap', x: 0, y: 0),
           ),
         ],
       );
@@ -295,7 +255,7 @@ void main() {
           MinorNation(
             id: 'm1',
             capitalProvinceId: 'oldWorld|mcap',
-            capitalTile: _tile('oldWorld|mcap', 0, 0),
+            capitalTile: _tile('oldWorld|mcap', x: 0, y: 0),
           ),
         ],
       );
